+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   task list
 - Change order of the legal tasks in the Transfer task list
 - The 'Team projects' section is now 'Your team projects'
+- The "School" contact type is now "School or academy"
 
 ## [Release 39][release-39]